adds image transform for size correct, fixes resnet (closes #4)
